lab10/requirements.txt
flas
psycopg2-binary